hotel_booking_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
kafka-python
